requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
xlrd==2.0.1
djangorestframework-simplejwt==5.4.0
